refactor: move filter component to separate file
